hw4/requirements.txt
Django
psycopg2>=2.9,<3.0
python-dotenv==0.19.0
djangorestframework==3.15.1
djangorestframework-simplejwt
django-cors-headers
flake8
